+++ src/com/intellij/idea/plugin/hybris/codeInsight/daemon/TSInterceptorSpringBuilderFactory.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -73,7 +73,7 @@ object TSInterceptorSpringBuilderFactory {
 
     fun createGutterBuilder(project: Project, typeCode: String): SpringGutterIconBuilder<SpringBeanPointer<*>>? {
         val clazz = JavaPsiFacade.getInstance(project)
-            .findClass(HybrisConstants.CLASS_INTERCEPTOR_MAPPING, GlobalSearchScope.allScope(project))
+            .findClass(HybrisConstants.CLASS_FQN_INTERCEPTOR_MAPPING, GlobalSearchScope.allScope(project))
             ?: return null
 
         val builder = SpringGutterIconBuilder.createBuilder(