+++ src/com/intellij/idea/plugin/hybris/flexibleSearch/injection/FlexibleSearchInjector.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -18,7 +18,6 @@
 package com.intellij.idea.plugin.hybris.flexibleSearch.injection
 
 import com.intellij.idea.plugin.hybris.flexibleSearch.injection.impl.FlexibleSearchToImpexInjectorProvider
-import com.intellij.idea.plugin.hybris.flexibleSearch.injection.impl.FlexibleSearchToJavaInjectorProvider
 import com.intellij.idea.plugin.hybris.flexibleSearch.injection.impl.FlexibleSearchToKotlinInjectorProvider
 import com.intellij.psi.InjectedLanguagePlaces
 import com.intellij.psi.LanguageInjector
@@ -32,8 +31,6 @@ class FlexibleSearchInjector : LanguageInjector {
     ) {
         FlexibleSearchToImpexInjectorProvider.instance
             ?.inject(host, injectionPlacesRegistrar)
-            ?: FlexibleSearchToJavaInjectorProvider.instance
-                ?.inject(host, injectionPlacesRegistrar)
             ?: FlexibleSearchToKotlinInjectorProvider.instance
                 ?.inject(host, injectionPlacesRegistrar)
     }