+++ src/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexSubTypeNameMixin.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -47,7 +47,7 @@ abstract class ImpexSubTypeNameMixin(node: ASTNode) : ASTWrapperReferencePsiElem
                         ?.attributeList
                         ?.firstOrNull()
                         ?.anyAttributeValue
-                        ?.text != HybrisConstants.CLASS_CONFIG_IMPORT_PROCESSOR
+                        ?.text != HybrisConstants.CLASS_FQN_CONFIG_IMPORT_PROCESSOR
             } != null
     ) {
         ImpexTSSubTypeItemReference(this)