+++ src/com/intellij/idea/plugin/hybris/polyglotQuery/injection/PolyglotQueryInjector.kt
@@ -18,7 +18,6 @@
 package com.intellij.idea.plugin.hybris.polyglotQuery.injection
 
 import com.intellij.idea.plugin.hybris.polyglotQuery.injection.impl.PolyglotQueryToImpexInjectorProvider
-import com.intellij.idea.plugin.hybris.polyglotQuery.injection.impl.PolyglotQueryToJavaInjectorProvider
 import com.intellij.idea.plugin.hybris.polyglotQuery.injection.impl.PolyglotQueryToKotlinInjectorProvider
 import com.intellij.psi.InjectedLanguagePlaces
 import com.intellij.psi.LanguageInjector
@@ -32,8 +31,6 @@ class PolyglotQueryInjector : LanguageInjector {
     ) {
         PolyglotQueryToImpexInjectorProvider.instance
             ?.inject(host, injectionPlacesRegistrar)
-            ?: PolyglotQueryToJavaInjectorProvider.instance
-                ?.inject(host, injectionPlacesRegistrar)
             ?: PolyglotQueryToKotlinInjectorProvider.instance
                 ?.inject(host, injectionPlacesRegistrar)
     }